requirements.txt
numpy==1.19.5
Pillow==9.0.1
scipy==1.6.2
torch==1.8.1
torchvision==0.9.1
